Crate ink_macro

Macros

blake2x256

Computes and expands into the BLAKE2b 256-bit hash of the string input.

selector_bytes

Computes the ink! selector of the string and expands into its byte representation.

selector_id

Computes the ink! selector of the string and expands into its u32 representation.

Attribute Macros

chain_extension

Defines the interface for a chain extension.

contract

Entry point for writing ink! smart contracts.

storage_item

Prepares the type to be fully compatible and usable with the storage. It implements all necessary traits and calculates the storage key for types. Packed types don’t have a storage key, but non-packed types (like Mapping, Lazy etc.) require calculating the storage key during compilation.

test

Defines a unit test that makes use of ink!’s off-chain testing capabilities.

trait_definition

Marks trait definitions to ink! as special ink! trait definitions.

Derive Macros

Storable

Derives ink::storage’s Storable trait for the given struct, enum or union.

StorableHint

Derives ink::storage’s StorableHint trait for the given struct or enum.

StorageKey

Derives ink::storage’s StorageKey trait for the given struct or enum.

StorageLayout

Derives ink::storage’s StorageLayout trait for the given struct or enum.